Move "default" connection settings to UI config
The values are set based on the last connection attempt. Additionally,
the options to enable direct MIDI input has been moved to the UI config,
so that the only things remaining in the INI config are the UD config
page assignments and the default UD3 features. Both of these are "very
advanced" options, so editing the config by hand is no longer necessary.