Indicate progress while transmitting VMS
As a side effect this commit changes VMS transmission speed to be based
on UD3 connection speed/quality. Specifically, a new frame will be send
as soon as the ACK for the previous frame is received, but no earlier.
This should make sure that we never overload the connection or the UD3.